requirements.txt
python-telegram-bot==13.7
pymongo==3.12.0
requests==2.26.0